requirements.txt
streamlit
google-api-python-client
azure-storage-blob
google-auth-oauthlib
google-auth  
google-auth-httplib2
bcrypt

